+++ macros/src/attr/struct.rs
@@ -1,10 +1,10 @@
 use std::collections::HashMap;
 
-use syn::{parse_quote, Attribute, Fields, Ident, Path, Result, Type, WherePredicate};
+use syn::{parse_quote, Attribute, Expr, Fields, Ident, Path, Result, Type, WherePredicate};
 
 use super::{
-    parse_assign_from_str, parse_assign_inflection, parse_bound, parse_concrete, parse_optional,
-    Attr, ContainerAttr, Optional, Serde, Tagged,
+    parse_assign_expr, parse_assign_from_str, parse_assign_inflection, parse_bound, parse_concrete,
+    parse_optional, Attr, ContainerAttr, Optional, Serde, Tagged,
 };
 use crate::{
     attr::{parse_assign_str, EnumAttr, Inflection, VariantAttr},
@@ -17,8 +17,8 @@ pub struct StructAttr {
     pub type_as: Option<Type>,
     pub type_override: Option<String>,
     pub rename_all: Option<Inflection>,
-    pub rename: Option<String>,
-    pub export_to: Option<String>,
+    pub rename: Option<Expr>,
+    pub export_to: Option<Expr>,
     pub export: bool,
     pub tag: Option<String>,
     pub docs: String,
@@ -159,11 +159,11 @@ impl_parse! {
         "crate" => out.crate_rename = Some(parse_assign_from_str(input)?),
         "as" => out.type_as = Some(parse_assign_from_str(input)?),
         "type" => out.type_override = Some(parse_assign_str(input)?),
-        "rename" => out.rename = Some(parse_assign_str(input)?),
+        "rename" => out.rename = Some(parse_assign_expr(input)?),
         "rename_all" => out.rename_all = Some(parse_assign_inflection(input)?),
         "tag" => out.tag = Some(parse_assign_str(input)?),
         "export" => out.export = true,
-        "export_to" => out.export_to = Some(parse_assign_str(input)?),
+        "export_to" => out.export_to = Some(parse_assign_expr(input)?),
         "concrete" => out.concrete = parse_concrete(input)?,
         "bound" => out.bound = Some(parse_bound(input)?),
         "optional_fields" => out.optional_fields = parse_optional(input)?,
@@ -172,7 +172,7 @@ impl_parse! {
 
 impl_parse! {
     Serde<StructAttr>(input, out) {
-        "rename" => out.0.rename = Some(parse_assign_str(input)?),
+        "rename" => out.0.rename = Some(parse_assign_expr(input)?),
         "rename_all" => out.0.rename_all = Some(parse_assign_inflection(input)?),
         "tag" => out.0.tag = Some(parse_assign_str(input)?),
         "bound" => out.0.bound = Some(parse_bound(input)?),