+++ ts-rs/src/lib.rs
@@ -208,12 +208,13 @@ mod tokio;
 ///   Specifies where the type should be exported to. Defaults to `<name>.ts`.  
 ///   The path given to the `export_to` attribute is relative to the `TS_RS_EXPORT_DIR` environment variable,
 ///   or, if `TS_RS_EXPORT_DIR` is not set, to `./bindings`  
-///   If the provided path ends in a trailing `/`, it is interpreted as a directory.   
+///   If the provided path ends in a trailing `/`, it is interpreted as a directory.  
+///   This attribute also accepts arbitrary expressions.  
 ///   Note that you need to add the `export` attribute as well, in order to generate a test which exports the type.
 ///   <br/><br/>
 ///
 /// - **`#[ts(as = "..")]`**  
-///   Overrides the type used in Typescript, using the provided Rust type instead.
+///   Overrides the type used in Typescript, using the provided Rust type instead.  
 ///   This is useful when you have a custom serializer and deserializer and don't want to implement `TS` manually
 ///   <br/><br/>
 ///
@@ -223,7 +224,8 @@ mod tokio;
 ///   <br/><br/>
 ///
 /// - **`#[ts(rename = "..")]`**  
-///   Sets the typescript name of the generated type
+///   Sets the typescript name of the generated type.  
+///   Also accepts expressions, e.g `#[ts(rename = module_path!().rsplit_once("::").unwrap().1)]`.
 ///   <br/><br/>
 ///
 /// - **`#[ts(rename_all = "..")]`**  
@@ -363,7 +365,8 @@ mod tokio;
 /// ### enum variant attributes
 ///
 /// - **`#[ts(rename = "..")]`**  
-///   Renames this variant. To rename all variants of an enum, see the container attribute `#[ts(rename_all = "..")]`.
+///   Renames this variant. To rename all variants of an enum, see the container attribute `#[ts(rename_all = "..")]`.  
+///   This attribute also accepts expressions, e.g `#[ts(rename = module_path!().rsplit_once("::").unwrap().1)]`.
 ///   <br/><br/>
 ///
 /// - **`#[ts(skip)]`**  
@@ -580,7 +583,7 @@ pub trait TS {
     ///
     /// If `T` cannot be exported (e.g because it's a primitive type), this function will return
     /// `None`.
-    fn output_path() -> Option<&'static Path> {
+    fn output_path() -> Option<PathBuf> {
         None
     }
 
@@ -618,10 +621,10 @@ pub struct Dependency {
     pub type_id: TypeId,
     /// Name of the type in TypeScript
     pub ts_name: String,
-    /// Path to where the type would be exported. By default a filename is derived from the types
+    /// Path to where the type would be exported. By default, a filename is derived from the types
     /// name, which can be customized with `#[ts(export_to = "..")]`.  
     /// This path does _not_ include a base directory.
-    pub output_path: &'static Path,
+    pub output_path: PathBuf,
 }
 
 impl Dependency {
@@ -748,7 +751,7 @@ macro_rules! impl_shadow {
             }
             fn decl() -> String { <$s as $crate::TS>::decl() }
             fn decl_concrete() -> String { <$s as $crate::TS>::decl_concrete() }
-            fn output_path() -> Option<&'static std::path::Path> { <$s as $crate::TS>::output_path() }
+            fn output_path() -> Option<std::path::PathBuf> { <$s as $crate::TS>::output_path() }
         }
     };
 }